style: card 코드 포맷 수정
